requirements.txt
openai
python-dotenv
schedule
